# perf(web): optimize timer resolution